Move utils specific to WAR into WAR
